Split main panel class into separate ui classes
This refactor changes the main panel code from one big class with if-statements to individual classes with poll methods.
