+++ backend/preview_collections.py
@@ -406,5 +406,7 @@ def _get_search_term(pcoll_type, sett) -> str:
 
 def get_hg_icon(icon_name) -> int:
     icon_list = preview_collections["hg_icons"]
-
-    return icon_list[icon_name].icon_id
+    try:
+        return icon_list[icon_name].icon_id
+    except IndexError:
+        return None
