+++ user_interface/main_panel/create_new_ui.py
@@ -1,0 +1,68 @@
+import bpy
+from HumGen3D.human.human import Human
+
+from .main_panel_baseclass import MainPanelPart
+
+
+class HG_PT_CREATE(MainPanelPart, bpy.types.Panel):
+    bl_idname = "HG_PT_CREATE"
+
+    @classmethod
+    def poll(self, context):
+        return not Human.find(context.object)
+
+    def draw_header(self, context):
+        self.draw_panel_switch_header(self.layout, context.scene.HG3D)
+
+    def draw(self, context):
+        """UI that shows when no human is selected, with buttons for creating a
+        new human.
+
+        Shows a template icon view of all 'starting humans', a switch for male
+        and female genders and a pink button to add the selected human
+        """
+
+        box = self.layout.box()
+
+        col = box.column(align=True)
+        col.label(text="Select a starting human")
+        col.template_icon_view(
+            context.scene.HG3D.pcoll,
+            "humans",
+            show_labels=True,
+            scale=10,
+            scale_popup=6,
+        )
+
+        row = col.row()
+        row.scale_y = 2
+        row.prop(context.scene.HG3D, "gender", expand=True)
+
+        col = box.column()
+        col.scale_y = 2
+        col.alert = True
+        col.operator("hg3d.startcreation", icon="COMMUNITY", depress=True)
+
+        if "hg_batch_marker" in context.object:
+            self._draw_batch_marker_notification(col)
+
+    def _draw_batch_marker_notification(self, layout):
+        col = layout.column(align=True)
+        col.separator()
+        col.separator()
+        col.scale_y = 0.4
+        row = col.row()
+        row.alignment = "CENTER"
+        row.label(text="Go to the batch panel to", icon="INFO")
+        row = col.row()
+        row.alignment = "CENTER"
+        row.label(text="generate humans from this")
+        row = col.row()
+        row.alignment = "CENTER"
+        row.label(text="batch marker.")
+
+        col.separator()
+
+        row = col.row()
+        row.alignment = "CENTER"
+        row.label(text="(Switch at the top of the UI)")
